Move token provider into Firestore
