+++ src/lib.rs
@@ -6,5 +6,6 @@
 pub mod auth;
 pub mod error;
 pub mod firestore;
-pub mod token;
-pub(crate) mod utils;
+mod service_account;
+
+pub use service_account::ServiceAccount;
